add history plugin to swagger
