+++ tests/swagger/support.py
@@ -1,9 +1,10 @@
 import unittest
+import json
 
 from bravado_core.spec import Spec
 from bravado_core.resource import build_resources
-from bravado_core.response import OutgoingResponse
-from bravado_core.request import IncomingRequest
+from bravado_core.response import OutgoingResponse, validate_response
+from bravado_core.request import IncomingRequest, unmarshal_request
 
 from ..support import (BaseWebTest, MINIMALIST_BUCKET, MINIMALIST_GROUP,
                        MINIMALIST_COLLECTION, MINIMALIST_RECORD)
@@ -11,14 +12,21 @@ from ..support import (BaseWebTest, MINIMALIST_BUCKET, MINIMALIST_GROUP,
 
 class SwaggerTest(BaseWebTest, unittest.TestCase):
 
+    settings = {'includes': [
+        'kinto.plugins.history',
+    ]}
+
     @classmethod
     def setUpClass(cls):
         # FIXME: solve memory issues from generating the spec multiple times
-        app = BaseWebTest().make_app()
+        app = BaseWebTest().make_app(settings=cls.settings)
 
         cls.spec_dict = app.get('/swagger.json').json
         cls.spec = Spec.from_dict(cls.spec_dict)
         cls.resources = build_resources(cls.spec)
+
+    def get_app_settings(self, extras=None):
+        return self.settings
 
     def setUp(self):
         super(SwaggerTest, self).setUp()
@@ -72,3 +80,14 @@ class SwaggerTest(BaseWebTest, unittest.TestCase):
         resp.json = lambda: response.json
 
         return resp
+
+    def validate_request_call(self, op, **kargs):
+        params = unmarshal_request(self.request, op)
+        response = self.app.request(op.path_name.format(**params),
+                                    body=json.dumps(self.request.json()).encode(),
+                                    method=op.http_method.upper(),
+                                    headers=self.headers, **kargs)
+        schema = self.spec.deref(op.op_spec['responses'][str(response.status_code)])
+        casted_resp = self.cast_bravado_response(response)
+        validate_response(schema, op, casted_resp)
+        return response
